pyproject.toml
[tool.poetry]
name = "oc-da-python-p12"
version = "0.1.0"
description = "Développez une architecture back-end sécurisée avec Python et SQL"
authors = ["Grégory Lama <gregory.lama@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
coverage = "^7.4.4"
iniconfig = "^2.0.0"
packaging = "^24.0"
passlib = "^1.7.4"
pluggy = "^1.4.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
sqlalchemy-utils = "^0.41.2"
typing-extensions = "^4.11.0"
click = "^8.1.7"
rich = "^13.7.1"
pyjwt = "^2.8.0"
inquirer = "^3.2.4"
sentry-sdk = {extras = ["sqlalchemy"], version = "^2.0.1"}
ipython = "^8.24.0"
python-dateutil = "^2.9.0.post0"
kink = "^0.8.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
